One more size_t conversion.

+++ common/service/sms/gsmmulti.c
@@ -34,9 +34,9 @@ unsigned char GSM_MakeSMSIDFromTime(void)
 	return retval;
 }
 
-void GSM_Find_Free_Used_SMS2(GSM_Coding_Type Coding,GSM_SMSMessage SMS, int *UsedText, int *FreeText, int *FreeBytes)
+void GSM_Find_Free_Used_SMS2(GSM_Coding_Type Coding,GSM_SMSMessage SMS, size_t *UsedText, size_t *FreeText, size_t *FreeBytes)
 {
-	int UsedBytes;
+	size_t UsedBytes;
 
 	switch (Coding) {
 	case SMS_Coding_Default_No_Compression:
